+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed a bug causing the conversation list to be visible under the tab bar and the navigation bar.
 
 ## [1.0.0-alpha25] - 2022-11-24